# candidate-service/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Candidate Service API
  description: REST API for the Candidate Service
  version: "1.0"
servers:
- url: http://localhost:8099/api
  description: Generated server url
paths:
  /admin/kafka/topics:
    get:
      tags:
      - admin-controller
      operationId: listTopics
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                uniqueItems: true
    post:
      tags:
      - admin-controller
      operationId: createTopic
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: partitions
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 3
      - name: replication
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/candidates:
    get:
      tags:
      - candidate-controller
      operationId: getAllCandidates
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateDTO'
  /api/v1/candidates/{candidateId}:
    get:
      tags:
      - candidate-controller
      operationId: getCandidateById
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CandidateDTO'
  /api/v1/candidates/{candidateId}/work-histories:
    get:
      tags:
      - candidate-work-history-controller
      operationId: getWorkHistoriesForCandidate
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateWorkHistoryDTO'
  /api/v1/candidates/{candidateId}/work-histories/{workHistoryId}:
    get:
      tags:
      - candidate-work-history-controller
      operationId: getWorkHistory
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: workHistoryId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CandidateWorkHistoryDTO'
  /api/v1/candidates/{candidateId}/skills:
    get:
      tags:
      - candidate-skill-controller
      operationId: getSkillsForCandidate
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateSkillDTO'
  /api/v1/candidates/{candidateId}/skills/{skillId}:
    get:
      tags:
      - candidate-skill-controller
      operationId: getSkill
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: skillId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CandidateSkillDTO'
  /api/v1/candidates/{candidateId}/educations:
    get:
      tags:
      - candidate-education-controller
      operationId: getEducationsForCandidate
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateEducationDTO'
  /api/v1/candidates/{candidateId}/educations/{educationId}:
    get:
      tags:
      - candidate-education-controller
      operationId: getEducation
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: educationId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CandidateEducationDTO'
  /api/v1/candidates/{candidateId}/certifications:
    get:
      tags:
      - candidate-certification-controller
      operationId: getCertificationsForCandidate
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateCertificationDTO'
  /api/v1/candidates/{candidateId}/certifications/{certificationId}:
    get:
      tags:
      - candidate-certification-controller
      operationId: getCertification
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: certificationId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CandidateCertificationDTO'
  /admin/kafka/topics/{topic}/describe:
    get:
      tags:
      - admin-controller
      operationId: describeTopic
      parameters:
      - name: topic
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TopicDescription'
  /admin/kafka/topics/{name}/offsets:
    get:
      tags:
      - admin-controller
      operationId: getEndOffsets
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
  /admin/kafka/consumer-groups/{groupId}/offsets:
    get:
      tags:
      - admin-controller
      operationId: getConsumerGroupOffsets
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/OffsetAndMetadata'
  /admin/kafka/topics/{name}:
    delete:
      tags:
      - admin-controller
      operationId: deleteTopic
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    CandidateDTO:
      properties:
        candidateId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        nationality:
          type: string
    CandidateWorkHistoryDTO:
      properties:
        workHistoryId:
          type: integer
          format: int32
        candidateId:
          type: string
          format: uuid
        companyName:
          type: string
        positionTitle:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        responsibilities:
          type: string
    CandidateSkillDTO:
      properties:
        candidateId:
          type: string
          format: uuid
        skillId:
          type: integer
          format: int32
        proficiencyLevel:
          type: string
        yearsOfExperience:
          type: integer
          format: int32
    CandidateEducationDTO:
      properties:
        educationId:
          type: integer
          format: int32
        candidateId:
          type: string
          format: uuid
        institution:
          type: string
        degree:
          type: string
        fieldOfStudy:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        grade:
          type: string
        notes:
          type: string
    CandidateCertificationDTO:
      properties:
        certificationId:
          type: integer
          format: int32
        candidateId:
          type: string
          format: uuid
        certificateName:
          type: string
        issuedBy:
          type: string
        issueDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
    TopicDescription:
      properties:
        internal:
          type: boolean
    OffsetAndMetadata: {}